controller added

+++ src/utils/cloudinary.js
@@ -1,29 +1,38 @@
-import {vs as cloudinary} from 'cloudinary' 
+import {v2 as cloudinary} from 'cloudinary'
 import fs from 'fs'
 
 cloudinary.config({
-    cloud_name:  process.env.CLOUDINARY_CLOUD_NAME,
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
     api_key: process.env.CLOUDINARY_API_KEY,
-    api_secret:  process.env.CLOUDINARY_API_SECRET,
+    api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
-const uploadOnCludinary=async (localFilePath)=>{
+const uploadCloudinary=async(localFilePath)=>{
     try{
-        if(!localFilePath) return null
-        // upload the file on cloudinary
-        const respnse=await cloudinary.uploader.upload
-        (localFilePath,{
-            resource_type:"auto"
-        })
-        console.log("file is uploaded on cloudinary")
+        if(!localFilePath) return null;
+        const response=await cloudinary.uploader.upload(localFilePath,{resource_type:'auto'});
+        console.log('File uploaded to Cloudinary:',response.url)
+        return response;
     }catch(error){
-        fs.unlinkSync(localFilePath)
+        if(fs.existsSync(localFilePath)){
+            fs.unlinkSync(localFilePath);
+        }
+        console.log("Upload Failed:",error);
         return null;
     }
 }
 
-cloudinary.v2.uploader.upload("https://upload.wikimedia.org/wikipedia/commons/a/ae/Olympic_flag.jpg",
-{ public_id: "olympic_flag" },
-function (error, result) { console.log(result); });
+cloudinary.uploader.upload(
+    "https://upload.wikimedia.org/wikipedia/commons/a/ae/Olympic_flag.jpg",
+    {public_id:"olymic flag"},
+    function(error,result){
+        if(error){
+            console.error("Remote upload failed:", error);
+        }
+        else{
+            console.log("Remote upload success:", result.secure_url);
+        }
+    }
+)
 
-export {app}
+export {uploadCloudinary};
